% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMR_DATABASES.R
\name{AMR_DATABASE_pipeline}
\alias{AMR_DATABASE_pipeline}
\title{AMR_DATABASES.R
Generic AMR Profiler : SampleNo_contigs.fasta vs. ARG_ANNOT database}
\usage{
AMR_DATABASE_pipeline(Org_id, SampleNo, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{SampleNo}{Sample number (or list of sample numbers) associated with contig.fasta file}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Takes Organism, Sample Number and queries a contig.fasta file
}
\details{
GAS AMR loci:
GAS AMR loci
dfrF, dfrG, DHFR, ermA, ermB, ermT, folA, folP, gyrA, mefAE, msrD, parC, rpsJ, tetM, tetO
note that ermA == ermTR, ermC == ermT, mefA/E includes mefA and mefE; DHFR is the same as folA
}
