% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMM2.R
\name{buildSNVtable}
\alias{buildSNVtable}
\title{M1UK typing pipeline from WGS SNVPhyl M1UK Analysis
December 21 2023, Walter Demczuk & Shelley Peterson}
\usage{
buildSNVtable(SNVTable, SNPList, pattern_ref)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{variant}{M1UK or M1DK - which SNP list should be referenced}

\item{locus}{Sample number associated with contig.fasta file}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
M1UK typing pipeline from WGS SNVPhyl M1UK Analysis
December 21 2023, Walter Demczuk & Shelley Peterson
}
\details{
GAS emm1 M1UK typing:
Use the SNV_Table.csv generated from the SNVphyl Galaxy pipeline, with MGAS5005 (CP000017.2) as a mapping reference.
include a known M1UK as a reference so that all SNV positions will be included in Galaxy SNPhyl outputs.
This will make LabWareUpload_GAS_M1UK.csv in the local Output directory.
Output.csv has SampleNo, SNVprofile and Genotype
if SNV profile matches MGAS5005 reference = "Global" genotype
if SNV profile has all 27 different = "M1UK"
Otherwise = "Intermediate (Number of SNVs vs references)"
}
