% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMM.R
\name{EMM_pipeline}
\alias{EMM_pipeline}
\title{emm typing pipeline from WGS assemblies - October 30, 2020}
\usage{
EMM_pipeline(Org_id, SampleNo, locus, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{SampleNo}{Sample number associated with contig.fasta file}

\item{locus}{Sample number associated with contig.fasta file}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
emm typing pipeline from WGS assemblies - October 30, 2020
}
\details{
GAS emm typing:
Takes Organism, Sample Number, Locus, and a Variable at queries a contig.fasta file.
For the most current emm types go to:
https://ftp.cdc.gov/pub/infectious_diseases/biotech/tsemm/
download the alltrimmed.fta file
rename to "emm_trimmed.fasta"
and place in the allele_lkup_dna folder in WADE GAS/emm_R
then be sure to index it using MakeBlastDB
}
