% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDuplicates.R
\name{find_sample_duplicates}
\alias{find_sample_duplicates}
\title{Duplicate Samples Finder
February 5 2024, Walter Demczuk & Shelley Peterson}
\usage{
find_sample_duplicates(Org_id, curr_work_dir)
}
\arguments{
\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Duplicate Samples Finder
February 5 2024, Walter Demczuk & Shelley Peterson
}
\details{
Return duplicate samples
Will prompt for a Labware LineList Export .xslx file
Scans for duplicate samples from the same patient 
based on prov-age-sex-DOB-organism-results
Generates Duplicates.csv file which will show whether or not samples have
been assigned linked sample numbers.
}
