% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonFunctions.R
\name{getdirectory}
\alias{getdirectory}
\title{Common helper functions for WADE analysis
December 21 2023, Shelley Peterson}
\usage{
getdirectory(curr_work_dir, Org_id, Test_id)
}
\arguments{
\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}

\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{Test_id}{AMR, TOXINS, VIRULENCE, NGSTAR, use MASTER for 16S id}

\item{SampleNo}{Sample number associated with contig.fasta file}

\item{LocusID}{The locus to query, or enter "list" to use a list of alleles}

\item{directorylist}{the list generated from the getdirectory function}

\item{SampleList.df}{the list of samples being used}

\item{reflist}{the list generated from the refdirectory function}

\item{NumLoci}{the number of loci being analyzed}

\item{LocusList.df}{the list of loci being analyzed}

\item{Blast_evalue}{the evalue parameter for the Blast function}

\item{CurrLocus}{the current locus being analyzed}

\item{LocusFile}{the location for the reference locus file}

\item{blastoutput}{the BLAST output file to be parsed}
}
\value{
a dataframe or list to be fed into other WADE functions
}
\description{
Common helper functions for WADE analysis
December 21 2023, Shelley Peterson
}
\details{
These are helper functions to streamline the other WADE functions
}
