% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Makeblastdb.R
\name{Index_pipeline}
\alias{Index_pipeline}
\title{Index BLAST fasta datases of lookup data}
\usage{
Index_pipeline(Org_id, Test_id, LocusID, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}

\item{TestID}{Sample number associated with contig.fasta file}

\item{locus}{Locus_name to build blast database}
}
\value{
A table frame containing the results of the query
}
\description{
Takes Organism, Test, and Locus and create BLAST databases
}
