% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabwareUpload_GONO_AMR.R
\name{labware_gono_amr}
\alias{labware_gono_amr}
\title{Labware Upload Formatter for GONO AMR
July 3 2024, Walter Demczuk & Shelley Peterson
#' Run AMR first, then run the 23S allele counts, and then the NGSTAR-MLST analyses
Then run this analysis to combine data from AMR, 23S rRNA and NG-STAR
to prepare full amr profile to upload to LabWare.}
\usage{
labware_gono_amr(Org_id, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Labware Upload Formatter for GONO AMR
July 3 2024, Walter Demczuk & Shelley Peterson
#' Run AMR first, then run the 23S allele counts, and then the NGSTAR-MLST analyses
Then run this analysis to combine data from AMR, 23S rRNA and NG-STAR
to prepare full amr profile to upload to LabWare.
}
