% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabwareUpload_PNEUMO_AMR.R
\name{labware_pneumo_amr}
\alias{labware_pneumo_amr}
\title{Labware Upload Formatter for PNEUMO AMR
July 3 2024, Walter Demczuk & Shelley Peterson
Run AMR first, then run the 23S allele counts,
Then run this analysis to combine data from AMR, 23S rRNA
to prepare full amr profile to upload to LabWare.}
\usage{
labware_pneumo_amr(Org_id, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Labware Upload Formatter for PNEUMO AMR
July 3 2024, Walter Demczuk & Shelley Peterson
Run AMR first, then run the 23S allele counts,
Then run this analysis to combine data from AMR, 23S rRNA
to prepare full amr profile to upload to LabWare.
}
