% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabwareUpload_PNEUMO_VIRULENCE.R
\name{labware_pneumo_virulence}
\alias{labware_pneumo_virulence}
\title{run MasterBlastR first, then this script outputs csv to LabWareUpload_GAS_TOXINS.csv in Output folder}
\usage{
labware_pneumo_virulence(Org_id, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Then run this analysis to combine data the full amr profile to upload to LabWare.
}
