% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Makeblastdb.R
\name{MakeblastDB_pipeline}
\alias{MakeblastDB_pipeline}
\title{Makeblastdb.R indexes Blast fasta databases
July 3, 2024, Walter Demczuk & Shelley Peterson}
\usage{
MakeblastDB_pipeline(Org_id, Test_id, LocusID, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{Test_id}{Sample number associated with contig.fasta file}

\item{LocusID}{Locus_name to build blast database}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Index BLAST fasta datases of lookup data
}
\details{
Takes Organism, Test, and Locus and create BLAST databases
}
