% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASTER.R
\name{MASTER_pipeline}
\alias{MASTER_pipeline}
\title{Molecular typing pipeline for WGS assemblies
July 3 2024, Walter Demczuk & Shelley Peterson}
\usage{
MASTER_pipeline(Org_id, Test_id, SampleNo, LocusID, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{Test_id}{AMR, TOXINS, VIRULENCE, NGSTAR, use MASTER for 16S id}

\item{SampleNo}{Sample number associated with contig.fasta file}

\item{LocusID}{The locus to query, or enter "list" to use a list of alleles}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Takes Organism, Sample Number, Locus, to query a contig.fasta file
}
\details{
GAS AMR loci:

dfrF, dfrG, DHFR, ermA, ermB, ermT, folA, folP, gyrA, mefAE, msrD, parC, rpsJ, tetM, tetO
note that ermA == ermTR, ermC == ermT, mefA/E includes mefA and mefE; DHFR is the same as folA
ermB + = ERY-R/CLI-R; ermT = ERY-R/CLI-?; mefA + = ERY-R/CLI-S; ermA/TR = ERY-R/CLI-Ind.

PNEUMO AMR loci:
cat ermB, ermTR, folA, folP, gyrA, parC, mefA, mefE, msrD, tetM, tetO, pbp1a, pbp2b, pbp2x
note that ermA == ermTR

GAS VIRULENCE FACTORS:
covR => V128A
covS => I332V; E226G (1228 deletion)
rocA => V47A; R259Q; V333A; D396N; I404T; P405S
rpoB/rgg => V169I
nga => promoter mutations: A8G (K3R) or T13G (F5V) or C17T (A6V) (protein): F5V+A6V=pnga3
nga => D349G is the D330G mutation
cpA, fctA, fctB, lepA and srtC1 all pilin proteins - use only fctA (major pilin protein)
fbp54, grab, ideS_mac, mf_spd, speB, scpA, ska always positive?
hasA,B,C are the same operon, use only hasA
}
