% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabwareUpload_Metrics.R
\name{metrics}
\alias{metrics}
\title{Labware Upload Formatter for WGS Metrics
February 5 2024, Walter Demczuk & Shelley Peterson
Browse to raw downloaded .csv files to create the upload format table for Labware
Doesn't matter in what order files are selected
Will combine columns into Labware upload ready .csv}
\usage{
metrics(Org_id, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Labware Upload Formatter for WGS Metrics
February 5 2024, Walter Demczuk & Shelley Peterson
Browse to raw downloaded .csv files to create the upload format table for Labware
Doesn't matter in what order files are selected
Will combine columns into Labware upload ready .csv
}
