% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC_Compare.R
\name{MIC_compare}
\alias{MIC_compare}
\title{Detects potententially contaminated samples based on plate proximity
February 5, 2024, Shelley Peterson
Run this analysis to evaluate consistency of MIC results}
\usage{
MIC_compare(curr_work_dir)
}
\arguments{
\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Detects potententially contaminated samples based on plate proximity
February 5, 2024, Shelley Peterson
Run this analysis to evaluate consistency of MIC results
}
