% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGMAST_MLST.R
\name{NGMAST_pipeline}
\alias{NGMAST_pipeline}
\title{NG-MAST MLST pipeline for WGS assemblies}
\usage{
NGMAST_pipeline(Org_id, SampleNo, locus, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GONO}

\item{SampleNo}{Sample number (or "list" to use a list of sample numbers) associated with contig.fasta file}

\item{locus}{The locus to query, (or "list" to use a list of alleles)}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Takes Organism, Sample Number, Locus, and queries a contig.fasta file vs. the NG-MAST Blast database
}
