% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSTAR_MLST.R
\name{NGSTAR_MLST_pipeline}
\alias{NGSTAR_MLST_pipeline}
\title{NG-STAR MLST pipeline for WGS assemblies}
\usage{
NGSTAR_MLST_pipeline(Org_id, SampleNo, locus, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{SampleNo}{Sample number (or list of sample numbers) associated with contig.fasta file}

\item{locus}{The locus to query, or enter list to use a list of alleles}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Takes Organism, Sample Number, Locus, and a Variable at queries a contig.fasta file
}
