% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newick_tree_labels.R
\name{nwk_rename_samples}
\alias{nwk_rename_samples}
\title{Replace sequencing sample numbers in a newick tree file
December 21 2023, Walter Demczuk & Shelley Peterson}
\usage{
nwk_rename_samples(Org_id, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Replace sequencing sample numbers in a newick tree file
December 21 2023, Walter Demczuk & Shelley Peterson
}
\details{
Replace node labels in newick file:
Replace sequencing sample numbers in a newick tree file
with custom sample numbers in .csv files
with old number label in first column with name "Samples:
with new number label in second column with name "TreeLabel"
}
