% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newick_sort_order.R
\name{nwk_sort_order}
\alias{nwk_sort_order}
\title{Gets the sort order for samples from a newick tree file.
February 5 2024, Walter Demczuk & Shelley Peterson}
\usage{
nwk_sort_order(Org_id, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Browse to newick file and gets the sample id sort order.
}
