% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFDB.R
\docType{data}
\name{Org_id}
\alias{Org_id}
\title{VFDB.R
Generic Virulence Factor Profiler : SampleNo_contigs.fasta vs. VFDB database}
\format{
An object of class \code{character} of length 1.
}
\usage{
Org_id
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{SampleNo}{Sample number (or list of sample numbers) associated with contig.fasta file}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Takes Organism, Sample Number and queries a contig.fasta file
}
\details{
searches for virulence factors in genomic data
}
\keyword{datasets}
