% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proximity_Test.R
\name{proximity_test}
\alias{proximity_test}
\title{Detects potententially contaminated samples based on plate proximity
February 5 2024, Shelley Peterson
Run labware_GONO_AMR first, then run this analysis to evaluate sample proximity}
\usage{
proximity_test(Org_id, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Detects potententially contaminated samples based on plate proximity
February 5 2024, Shelley Peterson
Run labware_GONO_AMR first, then run this analysis to evaluate sample proximity
}
