% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Remove_Duplicates_NF_fasta.R
\name{remove_duplicate_fasta}
\alias{remove_duplicate_fasta}
\title{Removes duplicate fasta sequences from output_dna_notfound.fasta when using master_blastR.}
\usage{
remove_duplicate_fasta(Org_id, allele_id, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
Removes duplicate fasta sequences from output_dna_notfound.fasta when using master_blastR.
}
\details{
Remove Duplicate Sequences

Removes duplicate sequences from dna_notfound.fasta and numbers them starting at the number
entered in locus entry area.  These can be copied and pasted directly into the allele_lookup fastas.
}
