% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Remove_Duplicates_fasta.R
\name{remove_duplicates_fasta}
\alias{remove_duplicates_fasta}
\title{Fasta Duplicates Remover
December 21 2023, Shelley Peterson
Removes duplicate sequences from fasta file.}
\usage{
remove_duplicates_fasta(curr_work_dir)
}
\arguments{
\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A fasta file containing the results of the query
}
\description{
Fasta Duplicates Remover
December 21 2023, Shelley Peterson
Removes duplicate sequences from fasta file.
}
\details{
Remove Duplicate Sequences

Removes duplicate sequences from dna_notfound.fasta.
These can be copied and pasted directly into the allele_lookup fastas.
}
