% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rename_Contigs.R
\name{rename_contigs}
\alias{rename_contigs}
\title{Rename Assemblies and move to main contigs folder
December 21 2023, Walter Demczuk & Shelley Peterson}
\usage{
rename_contigs(Org_id)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}
}
\value{
A table frame containing the results of the query
}
\description{
Rename files in contigs_new to remove _ _ numbers from file names from Galaxy:
}
