% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rename_Submitted_Reads.R
\name{rename_fastqs}
\alias{rename_fastqs}
\title{Renames submitted WGS reads to NML LabWare numbers
October 8 2024, Shelley Peterson}
\usage{
rename_fastqs(x)
}
\arguments{
\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
Copied files with the new NML lab numbers, outputs samples which failed to copy into the R console
}
\description{
This script renames submitted PT reads to NML numbers for upload into IRIDA
}
\details{
Copies submitted raw reads from submitted sample no to LabWare No.
Reads in a list of new LabWare Nos. and submitted read file numbers from metadata.csv
}
