% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23SrRNA.R
\name{rRNA23S_pipeline}
\alias{rRNA23S_pipeline}
\title{23SrRNA.R
23S rRNA pipeline for WGS assemblies to determine number of mutated alleles}
\usage{
rRNA23S_pipeline(Org_id, SampleNo, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{SampleNo}{Sample number or "list" or "folder" of sample numbers associated with VCF file(s)}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
23SrRNA.R
23S rRNA pipeline for WGS assemblies to determine number of mutated alleles
}
\details{
23S rRNA pipeline for WGS assemblies to determine number of mutated alleles

E.coli: A2059G and C2611T

GONO:
Run SNP core pipeline with NCCP11945_23S4.fasta file: A2045G or C2597T
PNEUMO:
Run SNP core pipeline with 23S_R6.fasta file: A2061G or C2613T

Takes Organism, Sample Number, Locus, and a Variable at queries a contig.fasta file
#Parses 23s rRNA mutations from VCF files

ON GALAXY:

Alternative_allele_proporition = 0.1
min_coverage = 15
min_mean_mapping = 30
run_name = 23S
in FreeBayes step of phylogeny set ploidy = 4

Export VCF from Galaxy:
Under "Collection Tools" , use the tool "Export to Warehouse" on "Filter vcf on collection 210".
It will bundle them together and makes a history item.  They are in /Warehouse/Temporary/galaxy_exports/RANDOM_NAME.
Click the "eye" to view RANDOM_NAME. You can then copy them from there using the regular file explorer into:
            " folder
}
