% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateLookups.R
\name{update_lookups}
\alias{update_lookups}
\title{July 3, 2024, Shelley Peterson}
\usage{
update_lookups(Org_Test, curr_work_dir)
}
\arguments{
\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}

\item{Test_id}{Test for which databases need to be updated}
}
\value{
A table frame containing the results of the query
}
\description{
Downloads lookup table data from URLs
}
\details{
Takes Organism, and test to update lookup tables
}
