% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLST.R
\name{MLST_pipeline}
\alias{MLST_pipeline}
\title{MLST pipeline for WGS assemblies
July 3, 2024, Walter Demczuk & Shelley Peterson}
\usage{
MLST_pipeline(Org_id, Test_id, SampleNo, LocusID, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{SampleNo}{Sample number associated with contig.fasta file}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}

\item{locus}{The locus to query, or enter list to use a list of alleles}
}
\value{
A table frame containing the results of the query
}
\description{
Takes Organism, Sample Number, Locus, and a Variable at queries a contig.fasta file
}
\details{
MLST pipeline:
Takes Organism, Sample Number, Locus, and a Variable at queries a contig.fasta file vs downloaded pubMLST data.
}
