% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16SrRNA.R
\name{rRNA16S_pipeline}
\alias{rRNA16S_pipeline}
\title{16SrRNA Pipeline
December 21 2023, Walter Demczuk & Shelley Peterson
Run this analysis to combine data output MasterBlastR profile to upload to LabWare.}
\usage{
rRNA16S_pipeline(Org_id, curr_work_dir)
}
\arguments{
\item{Org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{curr_work_dir}{Start up directory from pipeline project to locate system file structure}
}
\value{
A table frame containing the results of the query
}
\description{
16SrRNA Pipeline
December 21 2023, Walter Demczuk & Shelley Peterson
Run this analysis to combine data output MasterBlastR profile to upload to LabWare.
}
